audio/curve/vecmath/track/frustum/camera: prototype the float-argument implicit-declaration call sites
The -requireprotos census flagged 20 callees with float-by-value
parameters reached through implicit int declarations - the class where
default argument promotion can really change codegen. 36 TUs now
include the owning header (curve.h, vecmath.h, sys/objects.h,
sfx_position_api.h, sfx_channel_volume_api.h, lightmap_api.h,
track_dolphin_api.h, track_bbox_api.h, camera_shake_api.h, frustum.h,
maketex_sequence_api.h, pad_api.h, intersect_hud_api.h,
intersect_api.h). Five Sfx_PlayAtPositionFromObject sites drop their
(int)obj argument pun; ARWArwing's int-typed obj param spells the
GameObject* conversion explicitly. Every float argument at these sites
is an f32 lvalue, so the f64 promotion was invisible in codegen: all
36 objects neutral in every linked section, relocation and symbol,
main.dol byte-identical.

+++ src/dlls/objects/225_WispBaddie/WispBaddie.c
@@ -12,6 +12,8 @@
 #include "main/frame_timing.h"
 #include "main/mm.h"
 #include "sys/objects.h"
+#include "main/curve.h"
+#include "main/audio/sfx_position_api.h"
 
 #define WISPBADDIE_HIT_VOLUME_SLOT 10
 
@@ -149,7 +151,7 @@ void WispBaddie_update(GameObject* obj) {
             state->flags = (u8)(flags & ~WISPBADDIE_FLAG_CHASE_PLAYER);
             state->flags = (u8)(state->flags | WISPBADDIE_FLAG_CHASE_LOCKOUT);
         }
-        Sfx_PlayAtPositionFromObject((int)obj, hitPosX, hitPosY, hitPosZ, SFXTRIG_robolaser16);
+        Sfx_PlayAtPositionFromObject(obj, hitPosX, hitPosY, hitPosZ, SFXTRIG_robolaser16);
     }
 
     particleMode = 4;
